+++ src/app/shared/components/slideshow/slideshow.component.ts
@@ -5,10 +5,11 @@
  * https://opensource.org/licenses/MIT
  */
 
-import { Component, EventEmitter, Input, OnDestroy, OnInit, Output, QueryList, ViewChildren } from '@angular/core';
+import { Component, EventEmitter, forwardRef, Input, OnDestroy, OnInit, Output, QueryList, ViewChildren } from '@angular/core';
+import { ControlValueAccessor, FormControl, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { Router } from '@angular/router';
-import { BehaviorSubject, interval, Observable, Subscription } from 'rxjs';
-import { delay, mapTo, tap } from 'rxjs/operators';
+import { BehaviorSubject, interval, merge, Observable, Subscription } from 'rxjs';
+import { debounceTime, delay, mapTo, tap } from 'rxjs/operators';
 import { fadeInOut } from 'src/animations/fadeInOut';
 import { Image } from 'src/models/entities/Image';
 
@@ -16,19 +17,28 @@ import { Image } from 'src/models/entities/Image';
   selector: 'app-slideshow',
   templateUrl: './slideshow.component.html',
   styleUrls: ['./slideshow.component.css'],
+  providers: [
+    {
+      provide: NG_VALUE_ACCESSOR,
+      multi: true,
+      useExisting: forwardRef(() => (SlideshowComponent))
+    }
+  ],
   animations: [
     fadeInOut(500)
   ]
 })
 export class SlideshowComponent
-  implements OnInit, OnDestroy {
+  implements OnInit, OnDestroy, ControlValueAccessor {
 
+  private touchedSubscriptions: Subscription[] = [];
+  private valueChangesSubscriptions: Subscription[] = [];
   private currentIndex = 0;
   private currentIndexSource = new BehaviorSubject(this.currentIndex);
   private autoRotateImagesSubscription: Subscription;
   private autoRotationInterval = 5000;
 
-  @Input() @Output() images: Image[] = [];
+  @Input() images: Image[] = [];
   @Input() autocycle = true;
   @Input() editable = false;
   @Input() showSlideSelectors = true;
@@ -37,7 +47,10 @@ export class SlideshowComponent
   @Input() slideHeight = 'auto';
   @Output() navigate = new EventEmitter<void>();
   @Output() add = new EventEmitter<void>();
+  @Output() touched = new EventEmitter<void>();
   currentIndex$ = this.currentIndexSource.asObservable();
+
+  formControl = new FormControl();
 
   constructor(
     private router: Router
@@ -50,7 +63,41 @@ export class SlideshowComponent
   }
 
   ngOnDestroy(): void {
+    for (const sub of [
+      ...this.valueChangesSubscriptions,
+      ...this.touchedSubscriptions]) {
+      sub.unsubscribe();
+    }
     this.stopAutoRotation();
+  }
+
+  onTouched(): void {
+    this.touched.emit();
+  }
+
+  writeValue(obj: any): void {
+    if (Array.isArray(obj)) {
+      this.formControl.setValue(obj);
+      this.images = obj;
+    }
+  }
+
+  registerOnChange(fn: (value: any) => void): void {
+    const sub = this.formControl.valueChanges.pipe(debounceTime(250), tap(fn)).subscribe();
+    this.valueChangesSubscriptions.push(sub);
+  }
+
+  registerOnTouched(fn: () => void): void {
+    const sub = merge(this.touched).pipe(tap(fn)).subscribe();
+    this.touchedSubscriptions.push(sub);
+  }
+
+  setDisabledState?(isDisabled: boolean): void {
+    if (isDisabled) {
+      this.formControl.disable({ emitEvent: false });
+    } else {
+      this.formControl.enable({ emitEvent: false });
+    }
   }
 
   onClickImage(img: Image) {